API BidHistory

+++ source/e-auction/src/main/java/com/e/auction/api/view/form/bidhistory/CreateBidHistoryForm.java
@@ -11,5 +11,5 @@ public class CreateBidHistoryForm {
     @NotNull(message = "bidderId cannot be null!!!")
     private Long bidderId;
     @NotNull(message = "bidPrice cannot be null!!!")
-    private Long bidPrice;
+    private Double bidPrice;
 }
